web: refactored request handlers into multiple modules
